refactor: BoardPageContent 게시글 목록 컴포넌트·훅·타입 분리
