+++ src/features/board/components/board/BoardPageContent.tsx
@@ -1,24 +1,15 @@
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 import { useNavigate } from "react-router";
 import { Chip, RoundButton, toast } from "@/components/ui";
 import { cn } from "@/lib/cn";
-import { RecruitCard, RecruitCardSkeleton } from "@/components";
-import { DORMITORY_LABEL, ROOM_SIZE_LABEL, ROOM_SIZE_MAX } from "@/constants";
 import { CAMPUS_API_MAP, ROOM_FILTER_API_MAP, ROOM_FILTERS } from "@/features/board/constants";
 import type { RoomFilter } from "@/features/board/constants";
 import { ChecklistRequiredDialog } from "@/features/onboarding/components";
 import { usePostList, useRecommendedPostList } from "@/features/board/hooks";
-import { formatRelativeTime } from "@/features/board/utils";
-import { useFadeInOnScroll } from "@/hooks";
+import type { BoardPageContentProps } from "@/features/board/types";
 import { CampusSelector } from "./CampusSelector";
-
-type BoardPageContentProps = {
-  aiRecommend: boolean;
-  isOnboardingCompleted: boolean;
-  onAiRecommendChange: (value: boolean) => void;
-  onPostClick?: (id: number) => void;
-  onWriteClick?: () => void;
-};
+import { PostList } from "./PostList";
+import { RecommendedPostList } from "./RecommendedPostList";
 
 function BoardPageContent({
   aiRecommend,
@@ -34,8 +25,6 @@ function BoardPageContent({
     description: string;
     highlight?: string;
   } | null>(null);
-  const sentinelRef = useRef<HTMLDivElement>(null);
-  const fadeInRef = useFadeInOnScroll<HTMLDivElement>();
 
   const roomSizeParam = roomFilter ? ROOM_FILTER_API_MAP[roomFilter] : undefined;
 
@@ -54,20 +43,6 @@ function BoardPageContent({
   } = useRecommendedPostList({ roomSize: roomSizeParam });
 
   const posts = data?.pages.flatMap((page) => page.content) ?? [];
-
-  useEffect(() => {
-    const el = sentinelRef.current;
-    if (!el) return;
-
-    const observer = new IntersectionObserver((entries) => {
-      if (entries[0]?.isIntersecting && hasNextPage && !isFetchingNextPage) {
-        fetchNextPage();
-      }
-    });
-
-    observer.observe(el);
-    return () => observer.disconnect();
-  }, [fetchNextPage, hasNextPage, isFetchingNextPage]);
 
   return (
     <>
@@ -123,75 +98,22 @@ function BoardPageContent({
 
         {/* 게시글 목록 */}
         {aiRecommend ? (
-          isRecommendedError ? (
-            <div className="flex flex-1 items-center justify-center">
-              <p className="typo-body2 text-text-caption">추천 게시글을 불러오지 못했어요</p>
-            </div>
-          ) : isRecommendedLoading ? (
-            <div className="flex flex-col gap-[10px]">
-              {Array.from({ length: 5 }).map((_, i) => (
-                <RecruitCardSkeleton key={i} />
-              ))}
-            </div>
-          ) : !recommendedPosts || recommendedPosts.length === 0 ? (
-            <div className="flex flex-1 items-center justify-center">
-              <p className="typo-body2 text-text-caption">해당하는 추천 게시글이 없어요</p>
-            </div>
-          ) : (
-            <div ref={fadeInRef} className="flex flex-col gap-[10px]">
-              {recommendedPosts.map((post) => (
-                <RecruitCard
-                  key={post.postId}
-                  title={post.title}
-                  description={post.description}
-                  currentMembers={post.currentMemberCount}
-                  maxMembers={post.totalMemberCount}
-                  dormitory={DORMITORY_LABEL[post.dormitory] ?? post.dormitory}
-                  roomType={ROOM_SIZE_LABEL[post.roomSize] ?? post.roomSize}
-                  timeAgo={formatRelativeTime(post.createdAt)}
-                  onClick={() => onPostClick?.(post.postId)}
-                />
-              ))}
-            </div>
-          )
-        ) : isError ? (
-          <div className="flex flex-1 items-center justify-center">
-            <p className="typo-body2 text-text-caption">게시글을 불러오지 못했어요</p>
-          </div>
-        ) : isFetching && posts.length === 0 ? (
-          <div className="flex flex-col gap-[10px]">
-            {Array.from({ length: 5 }).map((_, i) => (
-              <RecruitCardSkeleton key={i} />
-            ))}
-          </div>
-        ) : !isFetching && posts.length === 0 ? (
-          <div className="flex flex-1 items-center justify-center">
-            <p className="typo-body2 text-text-caption">해당하는 게시글이 없어요</p>
-          </div>
+          <RecommendedPostList
+            posts={recommendedPosts}
+            isLoading={isRecommendedLoading}
+            isError={isRecommendedError}
+            onPostClick={onPostClick}
+          />
         ) : (
-          <div ref={fadeInRef} className="flex flex-col gap-[10px]">
-            {posts.map((post) => {
-              const maxMembers = ROOM_SIZE_MAX[post.roomSize] ?? 0;
-              const currentMembers = maxMembers - post.recruitMemberCount;
-
-              return (
-                <RecruitCard
-                  key={post.postId}
-                  title={post.title}
-                  description={post.description}
-                  currentMembers={currentMembers}
-                  maxMembers={maxMembers}
-                  dormitory={DORMITORY_LABEL[post.dormitory] ?? post.dormitory}
-                  roomType={ROOM_SIZE_LABEL[post.roomSize] ?? post.roomSize}
-                  timeAgo={formatRelativeTime(post.createdAt)}
-                  onClick={() => onPostClick?.(post.postId)}
-                />
-              );
-            })}
-
-            {/* 무한 스크롤 감지 */}
-            <div ref={sentinelRef} />
-          </div>
+          <PostList
+            posts={posts}
+            isFetching={isFetching}
+            isError={isError}
+            fetchNextPage={fetchNextPage}
+            hasNextPage={hasNextPage}
+            isFetchingNextPage={isFetchingNextPage}
+            onPostClick={onPostClick}
+          />
         )}
       </div>
