+++ src/features/board/components/board/index.ts
@@ -1,3 +1,5 @@
 export { AiRecommendBottomSheet } from "./AiRecommendBottomSheet";
 export { BoardPageContent, type BoardPageContentProps } from "./BoardPageContent";
 export { CampusSelector, CAMPUSES } from "./CampusSelector";
+export { PostList } from "./PostList";
+export { RecommendedPostList } from "./RecommendedPostList";
